renovate python coverage check
the first thing I noticed was that addresses.py wasn't in the coverage
report.  I figured out quickly why it wasn't covered, but then wondered
if it was possible to find uncovered files that weren't imported at all
After asking this question:https://stackoverflow.com/questions/69590196/is-it-possible-to-include-files-in-the-python-coverage-report-that-werent-touch
I figured out it is easier with version 6
but then it turned out that it still won't look in directories that are
not packages, unless they are listed.  And since tools is just a directory
and not a package, I had to add it to the list.  Then I realised that it
skips venv automatically, and everything else in the omit list isn't a
package and isn't called... so I didn't need the omit list, and so the
ignore check didn't make sense for this anymore either - I just have to
hope that I remember to include future non package directories

the original problem was that the script that I was testing worked fine
with python2 - so the test passed even though I didn't call it with the
environment.. and then it didn't get added to the coverage report